🚚 rename @store-io -> socket-io.filestore
