+++ app/web/studio/src/services/energy.ts
@@ -9,11 +9,6 @@ import {
 import z, { string } from "zod"
 
 import type energySchema from "~/app/node/store/gen/energy.schema.json"
-import {
-  socketIndex,
-  socketSchema,
-  socketSync,
-} from "~/packages/@store-io/src/json-store-io.web"
 import { now } from "~/packages/anvl/src/id/now"
 import type { Json } from "~/packages/anvl/src/json"
 import type { JsonSchema } from "~/packages/anvl/src/json-schema/json-schema"
@@ -22,6 +17,11 @@ import {
   removeFromIndex,
 } from "~/packages/hamr/src/recoil-tools/recoil-index"
 import type { Transact } from "~/packages/hamr/src/recoil-tools/recoil-transaction-tools"
+import {
+  socketIndex,
+  socketSchema,
+  socketSync,
+} from "~/packages/socket-io.filestore/src/json-store-io.web"
 
 import { energyFeaturesState } from "./energy_reaction"
 import { socket } from "./socket"
