+++ packages/socket-io.filestore/src/json-store-io.node.ts
@@ -1,6 +1,6 @@
 import type { Socket, Server as WebSocketServer } from "socket.io"
 
-import type { JsonStoreOptions } from "~/packages/@store-io/src"
+import type { JsonStoreOptions } from "~/packages/socket-io.filestore/src"
 import type { Encapsulate } from "~/packages/anvl/src/function"
 import type { Json, JsonArr } from "~/packages/anvl/src/json"
