+++ packages/socket-io.filestore/src/read.ts
