+++ packages/socket-io.filestore/src/utils.ts
