+++ packages/socket-io.filestore/src/write.ts
